readme: link tracking issue and matrix room
I wanted to try Cosmic and ended up here after googling my way to the tracking issue. 

Thanks for taking the time to make this flake!

I thought this readme could lead back to the tracking issue for people to easily find what is going on and test Cosmic on NixOS. 

Same for the matrix room.